+++ src/main/java/project/shop/CalculatingTotalCost.java
@@ -1,5 +1,7 @@
 package project.shop;
 
+import java.math.BigDecimal;
+
 public interface CalculatingTotalCost {
-    public double calculateTotalCost();
+    public BigDecimal calculateTotalCost();
 }
